ENH: Added category names to the eventbrite sources (#207)
* ENH: Added category names to the eventbrite sources

* FIX: handle multi-day events without venue in ANS

This handles the case when there's a multi-day event (e.g. https://anshome.org/events/conservation-cafe/) but there isn't a location. But, since we depend on a match between the number of events and the number of locations we scrape, we can't just let the error pass silently. Otherwise we get an error later on. So we set the location to an empty string when it's missing and let the schema test later on filter it out.